refactor: AnohterItem.tsx 컴포넌트를 components 폴더로 이동
Related to: #8
