+++ app/events/page.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 
 import { Input } from '@nextui-org/react';
 import EventCard from '@/containers/events/EventCard';
-import { KakaoMap } from '@/components';
+import { KakaoMap } from '@/components/map/KakaoMap';
 import IconUpArrow from '@/public/svgs/ArrowUp';
 import IconDownArrow from '@/public/svgs/ArrowDown';
 import AccordionMenu from '@/containers/events/AccordionMenu';
