+++ components/goods/index.tsx
@@ -1,0 +1,3 @@
+import AnotherItem from './AnotherItem';
+
+export default AnotherItem;
