--- components/index.ts
@@ -1,1 +1,0 @@
-export { default as KakaoMap } from './map/KakaoMap/KakaoMap';
