+++ components/map/KakaoMap/index.ts
@@ -1,0 +1,1 @@
+export { default as KakaoMap } from './KakaoMap';
