+++ containers/goods/index.tsx
@@ -1,2 +1,2 @@
 export { default as Item } from './Item';
-export { default as AnotherItem } from './AnotherItem';
+export { default as AnotherItem } from '../../components/goods/AnotherItem';
