+++ containers/home/index.tsx
@@ -4,3 +4,4 @@ export { default as GoodsItem } from './goods';
 export { default as SocialItem } from './social';
 export { default as EventItem } from './events';
 export { default as PopularArtist } from './popularArtist';
+export { default as PopularGoods } from './popularGoods';
