+++ containers/home/popularGoods/PopularGoods.tsx
@@ -1,0 +1,5 @@
+import Slider from 'react-slick';
+
+export default function PopularGoods() {
+  return <div>Slider</div>;
+}
