+++ containers/home/popularGoods/index.tsx
@@ -1,0 +1,3 @@
+import PopularGoods from './PopularGoods';
+
+export default PopularGoods;
